requirements.txt
pygame==2.5.2
pytmx==3.32
pyscroll==2.31 
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
Pillow>=10.0.0